+++ src/pages/Admin/admin.css
@@ -1,0 +1,8 @@
+.tag-list {
+  display: flex;
+  gap: 40px;
+}
+
+.tag-item ul {
+  list-style: none;
+}
