+++ src/pages/FindMeme/FindMeme.js
@@ -1,12 +1,12 @@
 import { useState, useEffect } from "react";
-import api from "../contexts/api";
+import api from "../../contexts/api";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom";
 import { Link } from "react-router-dom";
 import "./findmeme.css";
 import { CgSearch } from "react-icons/cg";
 import { CgSearchFound } from "react-icons/cg";
-import useAuthStore from "../store/useAuthStore";
+import useAuthStore from "../../store/useAuthStore";
 
 // 날짜 포맷팅 함수
 const formatDate = (dateString) => {
