+++ src/pages/FindMeme/FindMemePost.js
@@ -2,8 +2,8 @@ import { useState, useRef, useMemo, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import ReactQuill from "react-quill";
 import "react-quill/dist/quill.snow.css";
-import TagSelector from "../components/TagSelector";
-import api from "../contexts/api";
+import TagSelector from "../../components/TagSelector";
+import api from "../../contexts/api";
 import axios from "axios";
 
 function FindMemePost() {
