+++ src/pages/MemeDetail/MemeDetail.js
@@ -7,9 +7,9 @@ import { FaFacebookSquare } from "react-icons/fa";
 import { FaLink } from "react-icons/fa";
 import { AiFillLike } from "react-icons/ai";
 import { GoHeart, GoHeartFill } from "react-icons/go";
-import MemeSwiper from "../components/MemeSwiper";
+import MemeSwiper from "../../components/MemeSwiper";
 import { toast } from "react-toastify";
-import api from "../contexts/api";
+import api from "../../contexts/api";
 import "./memedetail.css";
 
 function MemeDetail() {
